contrib/gateway/requirements.txt
typing
uvicorn
fastapi
uuid
paho-mqtt
toml
datetime
